refactor(templates): ✏️ Consistent capital letters in nav bar
If being used in a sentence, only the first word will start with a capital, if being used in another way, all words will start with a capital. Conjunctions in lists (like and) should not start with capitals.
e.g. "About This Project" is wrong because it is a sentence
"Privacy and Security" is a list, so "and" does not begin with a capital

Signed-off-by: Matthew Gray <44288823+mmoomocow@users.noreply.github.com>